v3_0 update (#31)
* add new grapher configuration

* add graph21

* add graph21 code

* install additional required packages

* debugging first trial run

* further debugging

* clear axes before update

* graphing every 3 minutes

* autostart graph21

* fix QC @staticmethod

* fix QC commenting class

* fix QC: pythonic variable names

* fix QC: consecutive if statements

* fix QC   @staticmethod

* reduce upload frequency

* reset time

* add global constants

* parameterise timings

* QC: fix E305 2 empty lines after class

* separate start of graphs

* text fix

* also stop graphs

* fix typo

* resize graph a bit

* grphlist

* grphlist

* 30 iso 25 hours

* chmod 744

* GRAPH actions last because they take longer

* parameterize timings

* remove link to graph21

* revert removal

* remove graph21
allow inclusion in v3

* unneeded packages removed

+++ libdaemon.py
@@ -121,8 +121,7 @@ class Daemon:
         time.sleep(0.1)
 
     except OSError as err:
-      if err.errno == 3:
-        if os.path.exists(self.pidfile):
+      if (err.errno == 3) and os.path.exists(self.pidfile):
           os.remove(self.pidfile)
       else:
         print(str(err))
@@ -135,7 +134,8 @@ class Daemon:
     self.stop()
     self.start()
 
-  def run(self):
+  @staticmethod
+  def run():
     """
     You should override this method when you subclass Daemon.
     It will be called after the process has been
